polish header
