+++ apps/build-onchain-apps/src/components/Header.tsx
@@ -3,11 +3,11 @@ import { Flex, IconButton, Theme, Tooltip } from '@radix-ui/themes';
 import Image from 'next/image';
 import { GitHubLogoIcon } from '@radix-ui/react-icons';
 import NextLink from 'next/link';
+import { AccountConnectButton } from '@/onchain/components';
 import logo from '../../public/logo.svg';
 import { classNames } from '../utils/classNames';
 import { ThemeToggle } from './ThemeToggle';
 import styles from './Header.module.css';
-import { AccountConnectButton } from '@/onchain/components';
 
 export type HeaderProps = {
   children?: ReactNode;
@@ -47,22 +47,26 @@ function Header({ children, ghost }: HeaderProps) {
       >
         <div className={styles.HeaderInner}>
           <Flex align="center" position="absolute" top="0" bottom="0" left="0" pl="4">
-            <NextLink href="/" passHref legacyBehavior>
+            <NextLink href="/" passHref>
               <Image src={logo} alt="Onchain Coffee App" />
             </NextLink>
           </Flex>
 
-          <Flex align="center" gap="5" position="absolute" top="0" bottom="0" right="0" pr="4">
+          <Flex align="center" gap="5" position="absolute" top="0" bottom="0" right="0" pl="4">
             {children}
             <AccountConnectButton />
 
-              <Tooltip className="radix-themes-custom-fonts" content="View GitHub ">
-                <IconButton asChild size="3" variant="ghost" color="gray">
-                  <a href="https://github.com/coinbase/build-onchain-apps/tree/main/apps/build-onchain-apps" target="_blank" aria-labelledby="View GitHub Button">
-                    <GitHubLogoIcon width="16" height="16" />
-                  </a>
-                </IconButton>
-              </Tooltip>
+            <Tooltip className="radix-themes-custom-fonts" content="View GitHub ">
+              <IconButton asChild size="3" variant="ghost" color="gray">
+                <a
+                  href="https://github.com/coinbase/build-onchain-apps/tree/main/apps/build-onchain-apps"
+                  target="_blank"
+                  aria-labelledby="View GitHub Button"
+                >
+                  <GitHubLogoIcon width="16" height="16" />
+                </a>
+              </IconButton>
+            </Tooltip>
 
             <ThemeToggle />
           </Flex>
